moved and renamed abstract class and changed fields to protected
